notebooks/tutorials/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
mediapipe>=0.10.0
ultralytics>=8.0.0
qrcode>=7.4.0
Pillow>=10.0.0
streamlit-webrtc>=0.47.0
av>=10.0.0
google-genai>=0.3.0
google-generativeai>=0.3.0
